requirements.txt
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
